FA-23 - test(General): Improve golden tests with alchemist package
