+++ packages/ui_library/pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
     sdk: flutter
   golden_toolkit: ^0.13.0
   shared_preferences: ^2.0.13
+  alchemist: ^0.3.2
 
 dev_dependencies:
   flutter_test:
