+++ packages/ui_library/test/flutter_test_config.dart
@@ -1,10 +1,20 @@
 import 'dart:async';
 
-import 'package:flutter_test/flutter_test.dart';
+import 'package:alchemist/alchemist.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 
 Future<void> testExecutable(FutureOr<void> Function() testMain) async {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  const isRunningInCi = bool.fromEnvironment('CI', defaultValue: false);
   await loadAppFonts();
-  return testMain();
+
+  return AlchemistConfig.runWithConfig(
+    config: const AlchemistConfig(
+      platformGoldensConfig: PlatformGoldensConfig(
+        enabled: !isRunningInCi,
+      ),
+    ),
+    run: () {
+      return testMain();
+    },
+  );
 }
